simplified code for clearing the clipboard
