+++ app/src/main/AndroidManifest.xml
@@ -108,11 +108,6 @@
       android:windowSoftInputMode="adjustResize" />
 
     <service
-      android:name=".util.services.ClipboardService"
-      android:exported="false"
-      android:foregroundServiceType="shortService"
-      android:process=":clipboard_service_process" />
-    <service
       android:name=".util.services.PasswordExportService"
       android:exported="false"
       android:foregroundServiceType="shortService"
