+++ app/src/main/java/app/passwordstore/ui/crypto/PasswordCreationActivity.kt
@@ -55,6 +55,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import java.io.ByteArrayOutputStream
 import java.io.IOException
 import java.nio.file.Paths
+import java.util.concurrent.ScheduledExecutorService
 import javax.inject.Inject
 import kotlin.io.path.absolutePathString
 import kotlin.io.path.createDirectories
@@ -65,7 +66,6 @@ import kotlin.io.path.nameWithoutExtension
 import kotlin.io.path.pathString
 import kotlin.io.path.relativeTo
 import kotlin.io.path.writeBytes
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import logcat.LogPriority.ERROR
@@ -89,7 +89,7 @@ class PasswordCreationActivity : BasePGPActivity() {
   private var oldCategory: String? = null
   private var copy: Boolean = false
 
-  private var timer: Job? = null
+  private var timer: ScheduledExecutorService? = null
 
   private val otpImportAction =
     registerForActivityResult(StartActivityForResult()) { result ->
@@ -265,12 +265,12 @@ class PasswordCreationActivity : BasePGPActivity() {
   }
 
   override fun onResume() {
-    timer?.cancel()
+    timer?.shutdownNow()
     super.onResume()
   }
 
   override fun onPause() {
-    timer?.cancel()
+    timer?.shutdownNow()
     timer = startAutoDismissTimer()
     super.onPause()
   }
@@ -357,7 +357,8 @@ class PasswordCreationActivity : BasePGPActivity() {
       }
 
       if (copy) {
-        copyPasswordToClipboard(editPass)
+        clearTimer?.shutdownNow()
+        clearTimer = copyPasswordToClipboard(editPass)
       }
 
       // pass enters the key ID into `.gpg-id`.
