+++ app/src/main/res/values-ko-rKR/strings.xml
@@ -229,7 +229,6 @@
   <string name="button_create">생성</string>
   <string name="pref_search_on_start_title">시작 시 검색 열기</string>
   <string name="pref_search_on_start_summary">앱 실행 시 검색창 열기</string>
-  <string name="tap_clear_clipboard">클립보드를 지우려면 여기를 탭하세요</string>
   <string name="clone_git_repo">변경 사항을 동기화하기 전에 리포지토리를 복제해야 합니다.</string>
   <string name="pref_app_theme_title">앱 테마</string>
   <string name="pref_app_theme_value_light">밝은 테마</string>
