+++ app/src/main/res/values-pl-rPL/strings.xml
@@ -227,7 +227,6 @@
   <string name="button_create">Utwórz</string>
   <string name="pref_search_on_start_title">Otwórz wyszukiwanie przy starcie</string>
   <string name="pref_search_on_start_summary">Otwórz wyszukiwanie po uruchomieniu aplikacji</string>
-  <string name="tap_clear_clipboard">Dotknij tutaj, aby wyczyścić schowek</string>
   <string name="clone_git_repo">Repozytorium musi być sklonowane przed synchronizacją zmian.</string>
   <string name="pref_app_theme_title">Motyw aplikacji</string>
   <string name="pref_app_theme_value_light">Jasny</string>
