+++ app/src/main/res/values/strings.xml
@@ -261,7 +261,6 @@
   <string name="button_create">Create</string>
   <string name="pref_search_on_start_title">Open search on start</string>
   <string name="pref_search_on_start_summary">Open search bar when app is launched</string>
-  <string name="tap_clear_clipboard">Tap here to clear clipboard</string>
   <string name="clone_git_repo">The repository must be cloned before syncing changes.</string>
   <string name="pref_app_theme_title">App theme</string>
   <string name="pref_app_theme_value_light">Light</string>
